+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.3 - Released [not yet]
 
  * Fixed: when the shadow was enabled, the line-filling also had a shadow making it darker than it should be.
- * Added: new option: `options['gridDotted']` to make the grid lines dotted.
+ * Added: Issue #5: new option: `options['gridDotted']` to make the grid lines dotted.
  * Added: new option: `options['axisSize']` to specify the axis line size. (value 'auto' to use Grid setting)
  * Added: new option: `options['axisColor']` to specify the axis line color. (value 'auto' to use Grid setting)
  * Added: new option: `options['axisShadow']` to give the axis line a shadow. (value 'auto' to use Grid setting)
@@ -11,6 +11,7 @@
  * Fixed: `options['spacing']` now works as it should with the height instead of changing values.
  * Changed: the default style options have been changed a bit.
  * Changed: tabs in graph.js source to 8 spaces instead of 4, to see if Github formats it better then.
+ * Fixed: Issue #6: if you have `options['lineCurve']` enabled, but less than 3 data points it now disables lineCurve.
 
 ## v0.2 - Released May 18th, 2011